+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 本项目所有显著变更记录于此。格式遵循 [Keep a Changelog](https://keepachangelog.com/zh-CN/1.1.0/)，版本遵循[语义化版本](https://semver.org/lang/zh-CN/)。
 
-## [未发布]
+## [3.1.0] - 2026-09-10
 
 ### 新增